Звуки немножко попереносил. Плюс реформат кода.

+++ gameserver/src/main/java/net/sf/l2j/gameserver/handler/admincommandhandlers/AdminDelete.java
@@ -26,50 +26,40 @@ import net.sf.l2j.gameserver.network.SystemMessageId;
 /**
  * This class handles following admin commands: - delete = deletes target
  */
-public class AdminDelete implements IAdminCommandHandler
-{
-	private static final String[] ADMIN_COMMANDS =
-	{
-		"admin_delete"
-	};
-	
-	@Override
-	public boolean useAdminCommand(String command, L2PcInstance activeChar)
-	{
-		if (command.equals("admin_delete"))
-			handleDelete(activeChar);
-		
-		return true;
-	}
-	
-	@Override
-	public String[] getAdminCommandList()
-	{
-		return ADMIN_COMMANDS;
-	}
-	
-	private static void handleDelete(L2PcInstance activeChar)
-	{
-		L2Object obj = activeChar.getTarget();
-		if (obj != null && obj instanceof L2Npc)
-		{
-			L2Npc target = (L2Npc) obj;
-			target.deleteMe();
-			
-			L2Spawn spawn = target.getSpawn();
-			if (spawn != null)
-			{
-				spawn.stopRespawn();
-				
-				if (RaidBossSpawnManager.getInstance().isDefined(spawn.getNpcId()))
-					RaidBossSpawnManager.getInstance().deleteSpawn(spawn, true);
-				else
-					SpawnTable.getInstance().deleteSpawn(spawn, true);
-			}
-			
-			activeChar.sendMessage("Deleted " + target.getName() + " from " + target.getObjectId() + ".");
-		}
-		else
-			activeChar.sendPacket(SystemMessageId.INCORRECT_TARGET);
-	}
+public class AdminDelete implements IAdminCommandHandler {
+    private static final String[] ADMIN_COMMANDS =
+            {
+                    "admin_delete"
+            };
+
+    @Override
+    public boolean useAdminCommand(String command, L2PcInstance activeChar) {
+        if (command.equals("admin_delete")) { handleDelete(activeChar); }
+
+        return true;
+    }
+
+    @Override
+    public String[] getAdminCommandList() {
+        return ADMIN_COMMANDS;
+    }
+
+    private static void handleDelete(L2PcInstance activeChar) {
+        L2Object obj = activeChar.getTarget();
+        if (obj != null && obj instanceof L2Npc) {
+            L2Npc target = (L2Npc) obj;
+            target.deleteMe();
+
+            L2Spawn spawn = target.getSpawn();
+            if (spawn != null) {
+                spawn.stopRespawn();
+
+                if (RaidBossSpawnManager.getInstance().isDefined(spawn.getNpcId())) { RaidBossSpawnManager.getInstance().deleteSpawn(spawn, true); }
+                else { SpawnTable.getInstance().deleteSpawn(spawn, true); }
+            }
+
+            activeChar.sendMessage("Deleted " + target.getName() + " from " + target.getObjectId() + ".");
+        }
+        else { activeChar.sendPacket(SystemMessageId.INCORRECT_TARGET); }
+    }
 }
